Fix swinir arch name
